requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data Validation
pydantic==2.10.6
pydantic-settings==2.7.1

# AI Integration
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==8.3.5
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-mock==3.14.1

# Optional (for future enhancements)
# sqlmodel==0.0.14
# supabase==2.3.0

